remove some duplicate encoding configs
