+++ app/evmante/evmante_handler_test.go
@@ -51,7 +51,7 @@ func (s *TestSuite) TestAnteHandlerEVM() {
 			},
 			txSetup: func(deps *evmtest.TestDeps) sdk.FeeTx {
 				txMsg := evmtest.HappyTransferTx(deps, 0)
-				txBuilder := deps.EncCfg.TxConfig.NewTxBuilder()
+				txBuilder := deps.App.GetTxConfig().NewTxBuilder()
 
 				gethSigner := gethcore.LatestSignerForChainID(deps.App.EvmKeeper.EthChainID(deps.Ctx))
 				err := txMsg.Sign(gethSigner, deps.Sender.KeyringSigner)
@@ -77,7 +77,7 @@ func (s *TestSuite) TestAnteHandlerEVM() {
 						AccountKeeper:          deps.App.AccountKeeper,
 						BankKeeper:             deps.App.BankKeeper,
 						FeegrantKeeper:         deps.App.FeeGrantKeeper,
-						SignModeHandler:        deps.EncCfg.TxConfig.SignModeHandler(),
+						SignModeHandler:        deps.App.GetTxConfig().SignModeHandler(),
 						SigGasConsumer:         authante.DefaultSigVerificationGasConsumer,
 						ExtensionOptionChecker: func(*codectypes.Any) bool { return true },
 					},
