+++ x/devgas/v1/module_test.go
@@ -3,7 +3,7 @@ package devgas_test
 import (
 	abci "github.com/cometbft/cometbft/abci/types"
 
-	"github.com/NibiruChain/nibiru/v2/app/codec"
+	"github.com/NibiruChain/nibiru/v2/app"
 	devgas "github.com/NibiruChain/nibiru/v2/x/devgas/v1"
 	devgastypes "github.com/NibiruChain/nibiru/v2/x/devgas/v1/types"
 )
@@ -31,7 +31,7 @@ func (s *GenesisTestSuite) TestAppModule() {
 		s.EqualValues(s.genesis, *genState)
 
 		s.T().Log("AppModuleBasic.ValidateGenesis")
-		encCfg := codec.MakeEncodingConfig()
+		encCfg := app.MakeEncodingConfig()
 		err = appModule.AppModuleBasic.ValidateGenesis(cdc, encCfg.TxConfig, jsonBz)
 		s.NoError(err)
