+++ x/evm/evmtest/test_deps.go
@@ -9,7 +9,6 @@ import (
 	"github.com/ethereum/go-ethereum/core/vm"
 
 	"github.com/NibiruChain/nibiru/v2/app"
-	"github.com/NibiruChain/nibiru/v2/app/codec"
 	"github.com/NibiruChain/nibiru/v2/eth"
 	"github.com/NibiruChain/nibiru/v2/x/common/testutil/testapp"
 	"github.com/NibiruChain/nibiru/v2/x/evm"
@@ -20,22 +19,18 @@ import (
 type TestDeps struct {
 	App       *app.NibiruApp
 	Ctx       sdk.Context
-	EncCfg    codec.EncodingConfig
 	EvmKeeper *keeper.Keeper
 	GenState  *evm.GenesisState
 	Sender    EthPrivKeyAcc
 }
 
 func NewTestDeps() TestDeps {
-	encCfg := app.MakeEncodingConfig()
-	evm.RegisterInterfaces(encCfg.InterfaceRegistry)
-	eth.RegisterInterfaces(encCfg.InterfaceRegistry)
 	app, ctx := testapp.NewNibiruTestAppAndContext()
 	ctx = ctx.WithChainID(eth.EIP155ChainID_Testnet)
+
 	return TestDeps{
 		App:       app,
 		Ctx:       ctx,
-		EncCfg:    encCfg,
 		EvmKeeper: app.EvmKeeper,
 		GenState:  evm.DefaultGenesisState(),
 		Sender:    NewEthPrivAcc(),
