+++ x/evm/keeper/msg_ethereum_tx_test.go
@@ -256,11 +256,11 @@ func (s *Suite) TestEthereumTx_ABCI() {
 	}.Build()
 	s.NoError(err)
 
-	txBuilder := deps.EncCfg.TxConfig.NewTxBuilder()
+	txBuilder := deps.App.GetTxConfig().NewTxBuilder()
 	blockTx, err := evmTxMsg.BuildTx(txBuilder, evm.EVMBankDenom)
 	s.Require().NoError(err)
 
-	txBz, err := deps.EncCfg.TxConfig.TxEncoder()(blockTx)
+	txBz, err := deps.App.GetTxConfig().TxEncoder()(blockTx)
 	s.Require().NoError(err)
 	deliverTxResp := deps.App.DeliverTx(abci.RequestDeliverTx{Tx: txBz})
 	s.Require().True(deliverTxResp.IsOK(), "%#v", deliverTxResp)
